+++ src/.internal/core/Registry.ts
@@ -6,7 +6,7 @@ export class Registry {
 	/**
 	 * Currently running version of amCharts.
 	 */
-	readonly version: string = "5.2.37";
+	readonly version: string = "5.2.38";
 
 	/**
 	 * List of applied licenses.